+++ src/user/pruning/RemainingVerticesPruning.java
@@ -12,8 +12,8 @@ public class RemainingVerticesPruning implements PruningInterface{
 		Object[] JCLvars = (Object[]) args[0];
 		@SuppressWarnings("unchecked")
 		ObjectSet<String> remainingAux = (ObjectSet<String>) args[2];
-		float upperBound = (float) args[9];
-		float currentUpperBound = (float) args[8];
+		double upperBound = (double) args[9];
+		double currentUpperBound = (double) args[8];
 		DeltaEvaluationInterface edgeCalc = (DeltaEvaluationInterface) args[10];
 		
 		for(String vertex:remainingAux){
